+++ game/src/network/MsgHandler.h
@@ -2,6 +2,7 @@
 #define MSGHANDLER_H_
 #include "common/Singleton.h"
 #include <iostream>
+
 class GameRole;
 class InnerHead;
 class MsgHandler :public Singleton<MsgHandler>
@@ -11,8 +12,8 @@ private:
 public:
     MsgHandler(/* args */){};
     ~MsgHandler(){};
-    static void  onLogin(int playerId, GameRole* GameRole);
-    static void  onSomeMsg(int playerId, InnerHead* InnerHead );
+    static void  onLogin(int64_t playerId, GameRole* GameRole);
+    static void  onSomeMsg(int64_t playerId, InnerHead* InnerHead );
 };
 
 #endif
